[Chore] Decouple Serializers
